+++ src/main/java/ar/edu/utn/frc/siga/allocation/service/command/AllocationItem.java
@@ -3,5 +3,5 @@ package ar.edu.utn.frc.siga.allocation.service.command;
 import org.springframework.modulith.NamedInterface;
 
 @NamedInterface("api")
-public record AllocationItem(AllocationTarget target, Integer classroomId) {
+public record AllocationItem(AllocationTarget target, Long classroomId) {
 }
