+++ src/main/java/ar/edu/utn/frc/siga/auth/repository/RefreshTokenRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Integer> {
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
 
     Optional<RefreshToken> findByTokenHash(String tokenHash);
 
@@ -25,7 +25,7 @@ public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Inte
     @Query("update RefreshToken r set r.revoked = true, r.revokedAt = :revokedAt, "
             + "r.revocationReason = ar.edu.utn.frc.siga.auth.model.RevocationReason.CASCADE "
             + "where r.user.id = :userId and r.revoked = false")
-    int revokeAllByUserId(@Param("userId") Integer userId, @Param("revokedAt") Instant revokedAt);
+    int revokeAllByUserId(@Param("userId") Long userId, @Param("revokedAt") Instant revokedAt);
 
     @Modifying(clearAutomatically = true, flushAutomatically = true)
     @Query("update RefreshToken r set r.revoked = true, r.revokedAt = :revokedAt, "
