+++ src/main/java/ar/edu/utn/frc/siga/common/util/RoomDate.java
@@ -2,5 +2,5 @@ package ar.edu.utn.frc.siga.common.util;
 
 import java.time.LocalDate;
 
-public record RoomDate(Integer classroomId, LocalDate date) {
+public record RoomDate(Long classroomId, LocalDate date) {
 }
