+++ src/main/java/ar/edu/utn/frc/siga/roomrequest/validator/RoomRequestValidator.java
@@ -100,7 +100,7 @@ public class RoomRequestValidator {
     }
 
     private void validateClassroomsExist(CreateRoomRequestDto dto) {
-        Set<Integer> classroomIds = new LinkedHashSet<>();
+        Set<Long> classroomIds = new LinkedHashSet<>();
         for (CreateRoomRequestItemDto item : dto.items()) {
             if (item.currentClassroomId() != null) {
                 classroomIds.add(item.currentClassroomId());
@@ -111,9 +111,9 @@ public class RoomRequestValidator {
             return;
         }
 
-        Map<Integer, ClassroomResponseDto> classroomsById =
+        Map<Long, ClassroomResponseDto> classroomsById =
                 Maps.byId(classroomService.findByIds(classroomIds), ClassroomResponseDto::id);
-        for (Integer classroomId : classroomIds) {
+        for (Long classroomId : classroomIds) {
             if (!classroomsById.containsKey(classroomId)) {
                 throw ResourceNotFoundException.of("Classroom", classroomId);
             }
